Rename back to lib/api
